Logistic, Ordinal, and Multiple Imputation Models
This .do file concerns itself with the modelling of the effect of type of school on the number of o-level passes (controlling for other variables). This code provided for successful testing of logit, proportional odds, continuation ratio, and multiple imputation models. The O-level variable would benefit from being treated like a count variable rather than an ordinal categorical one. Also to note, dealing with missing data was successful, but in the future, I need to focus on asking why that data is missing (some people died, others were not in LEA authority and that missing data was generated). One last improvement is to include more model testing- for example a Brant test.