Secure access .do File upload
These files pertain to the secure access linked NPD dataset and are used to replicate the BHPS+UKHLS sections of the PhD. Additionally, the pooled analysis had to be conducted in a secure setting also, this means that the pooled files are also uploaded. All Ns, mentions to strata, and any senstivity or disclosive information has been redacted.